requirements.txt
streamlit==0.86.0
Pillow==8.3.1